requirements.txt
numpy==1.24
scipy
pandas
rdkit
click
cirpy
scikit-learn
matplotlib
igraph
pubchempy
numba
torch
tqdm
igraph
louvain
pillow
nmrglue
git+https://github.com/miguelarbesu/nmrplot